[Issue 87] Refactor `Settings` container/component
