+++ src/components/Settings/index.js
@@ -1,1 +1,1 @@
-export { default } from './Settings';
+export { default } from './Settings.container';
